+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- A setting to ignore owners and privileges now works for libraries as well.
+
 ### Fixed
 
 - Fixed an error when formatting an invalid function.